# Conflicts:
#	docs/UserGuide.md

+++ src/main/java/seedu/flashnotes/logic/Logic.java
@@ -69,12 +69,12 @@ public interface Logic {
     /**
      * Returns the user prefs' review card limit.
      */
-    Integer getReviewCardLimit();
+    long getReviewCardLimit();
 
     /**
      * Set the user prefs' review card limit.
      */
-    void setReviewCardLimit(Integer reviewCardLimit);
+    void setReviewCardLimit(long reviewCardLimit);
 
     /**
      * Update the user's review score for deck used in review.
