+++ src/test/java/seedu/flashnotes/logic/commands/AddDeckCommandTest.java
@@ -248,12 +248,12 @@ public class AddDeckCommandTest {
         }
 
         @Override
-        public Integer getReviewCardLimit() {
+        public long getReviewCardLimit() {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
-        public void setReviewCardLimit(Integer reviewCardLimit) {
+        public void setReviewCardLimit(long reviewCardLimit) {
             throw new AssertionError("This method should not be called.");
         }
